requirements.txt
numpy==1.13.3
pandas==0.20.3
Django==2.0.5
django-behave==0.1.5
django-debug-toolbar==1.11
django-celery==3.1.17
django-secure==1.0
django-passwords==0.2.0
django-tastypie==0.11.1
djangorestframework==3.4.0
flatdict==1.2.0
isoweek==1.3.0
django-nose==1.4.4
py_expression_eval
gunicorn==19.3.0
html2text
markdown==2.6
multipledispatch==0.4.8
jsonfield==1.0.3
django-extensions
factory-boy==2.7.0
django-model-changes==0.15
django-pandas==0.2.2
six==1.10.0
psycopg2==2.7.3.2
numexpr==2.6.4
django-betterforms==1.1.2
django-admin-bootstrapped==2.5.6
django-colorful==1.1.0
django-jsonfield==0.9.15
django-json-field==0.5.7
preppy==2.3.4
requests==2.9.1
mock==1.3.0
deepdiff==3.3.0
celery==4.1.0
